requirements.txt
grpcio
grpcio-tools
numpy
scikit-learn